Melhorias de tema e funcionalidade em todo o App.

+++ lib/models/user_model.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 class UserModel {
   final String name;
   final String photoUrl;
-  final double score;
+  final int score;
 
   UserModel({required this.name, required this.photoUrl, this.score = 0});
